Add env variable OWNCLOUD_APPS_DEPRECATED
To disable deprecated apps before a version upgrade we are introducing
this new env variables. This way the admin can control which app gets
properly disabled before an upgrade gets applies.

This is especially usefull for upgrading 10.2.x to 10.3.x since gallery
and files_videoplayer won't get bundles with the core tarball anymore
and the latter one is also not published to the marketplace which could
result in a broken upgrade process. Add an env variable like
OWNCLOUD_APPS_DEPRECATED=gallery,files_videoplayer to the container
start and it will be disabled properly.