Split HTTP_X_FORWARDED_HOST for getting first domain
By definition the HTTP_X_FORWARDED_HOST header COULD include a list of
forwarded hostnames separated by a comma, and maybe even follwoed by a
whitespace, in that case so far we had a wrong domain name set. To
prevent that we are always splitting the string by a comma, and trim the
surrounding whitespaces to get the first matching hostname.