In the current implementation environment variables take precedence over runtime changes. That results in bad UX, was reported earlier and dismissed in https://github.com/owncloud/core/issues/38535
With loglevel, the situation is especially bad, as a default value of
"2" is set to the environment variables, which render te careful
tests in override-config.php ineffective. Thus runtime changes are
not possibl in a docker deployment, even without passing OWNCLOUD_LOGLEVEL
into the docker environment.

The expectation is, that the loglevel can be set at runtime via

* Admin Settings UI
* occ log:manage --level=XXXX
* editing config.php

Currently these are silently ignored in docker deployments.

This is a minimal one-liner fix, to no longer enforce a default of "2".
With this fix, UI, occ and config.php work normally unless OWNCLOUD_LOGLEVEL
was explicitly specified in the docker environment.

The complete fix would make the environment lower precedence than runtime settings.
(I could not find a quick solution for that, as config.php's CONFIG array is not accessible
from override-config.php -- so it can really only implement overrides, not fallbacks)